Run tests in Go 1.6.3
We currently build the release in Go 1.6.3, so we should make sure that
it works. Checking the Go tip (master branch) is not necessary until we
plan to move on to a next release. Builds from tip take a lot more time.
An example build with 1.6.3 took 25s, versus 1m52s of tip, because
Travis needed to compile a Go release just for us in the latter case.