Do not show some errors in stdout
We want a way to log errors to dump.log, while not bothering users with
common errors that do not reflect actual problems.

For instance, it is common that a call to fetch previous logs will fail
to find a previous pod.

Another example are commands that are available in OSE 3.3 but not in
OSE 3.2, and those are executed regardless of the current client version
in a best effort manner.