Add developer-only toggle to print all errors
When developing, we often want to see errors that we don't want in the
final output to a regular user.
Even though all errors are always logged to dump.log in the generated
archive, it is more convenient to see them straigh away in stderr,
without needing to go open the dump archive.